requirements.txt
tkinterweb
markdown
Pillow
